增加了所有的请求响应类型，现在在重写Handler部分

+++ wx_event.go
@@ -4,7 +4,7 @@ type (
 	EventRequest struct {
 		ToUserName   string
 		FromUserName string
-		CreateTime   int
+		CreateTime   int64
 		MsgType      string
 		Event        string
 	}
